refactor: adjust infra layer responsabilities.
