+++ src/main/factories/http/index.ts
@@ -1,2 +1,3 @@
 export * from "./api-url-factory";
 export * from "./axios-http-client-factory";
+export * from "./headers-factory";
